requirements.txt
python-telegram-bot
rebelai
iso639
